Move reference check to field navigation
