+++ field_friend/automations/navigation/navigation.py
@@ -8,23 +8,11 @@ import numpy as np
 import rosys
 from nicegui import ui
 from rosys.analysis import track
-from rosys.geometry import GeoReference
-from rosys.hardware import Gnss
 
 from ..implements.implement import Implement
 
 if TYPE_CHECKING:
     from ...system import System
-
-
-def is_reference_valid(gnss: Gnss, *, max_distance: float = 5000.0) -> bool:
-    if GeoReference.current is None:
-        return False
-    if gnss.last_measurement is None:
-        return False
-    if gnss.last_measurement.gps_quality == 0:
-        return False
-    return gnss.last_measurement.point.distance(GeoReference.current.origin) <= max_distance
 
 
 class WorkflowException(Exception):
@@ -61,8 +49,6 @@ class Navigation(rosys.persistence.PersistentModule):
             if not await self.prepare():
                 self.log.error('Preparation failed')
                 return
-            if not is_reference_valid(self.gnss):
-                raise WorkflowException('reference to far away from robot')
             self.start_position = self.robot_locator.pose.point
             if isinstance(self.driver.wheels, rosys.hardware.WheelsSimulation) and not rosys.is_test:
                 self.create_simulation()
